requirements.txt
kafka-python-ng==2.2.2
faker==24.0.0
pyspark==3.5.0
